// dashboard/src/client/models/run-snapshot-event.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Run } from './run';
// May contain unused imports in some cases
// @ts-ignore
import type { RunNodeStatus } from './run-node-status';

export interface RunSnapshotEvent {
    'kind': RunSnapshotEventKindEnum;
    'run': Run;
    'nodes'?: Array<RunNodeStatus> | null;
}

export const RunSnapshotEventKindEnum = {
    RunSnapshot: 'run.snapshot'
} as const;

export type RunSnapshotEventKindEnum = typeof RunSnapshotEventKindEnum[keyof typeof RunSnapshotEventKindEnum];